// packages/receipt/src/generated/types.ts
export type SoftwareOperationType =
  'AE' | 'ONLINE_ANDROID' | 'ONLINE_IOS' | 'ONLINE_HUAWEI' | 'ONLINE_SERVICE';

/**
 * Software data type
 *
 * A program adatai típus
 */
export interface SoftwareType {
  /**
   * Software ID
   *
   * A program azonosítója
   */
  softwareId: string;
  /**
   * Software name
   *
   * A program neve
   */
  softwareName: string;
  /**
   * Software operation method (AE-based (hardware) or online)
   *
   * A program működési típusa (AE alapú (hardveres) vagy online szolgáltatás)
   */
  softwareOperation: SoftwareOperationType;
  /**
   * Software main version
   *
   * A program főverziója
   */
  softwareMainVersion: string;
  /**
   * The verification code of the software calculated with the SHA256 algorithm
   *
   * A szoftver SHA256 algoritmussal kiszámított ellenőrző kódja
   */
  softwareHash: string;
  /**
   * Software last update time
   *
   * A szoftver legutóbbi frissítésének időpontja
   */
  softwareLastUpdateTime: string;
  /**
   * Name of the software's developer
   *
   * A program fejlesztőjének neve
   */
  softwareDevName: string;
  /**
   * Electronic contact of the software's developer
   *
   * A program fejlesztőjének elektronikus elérhetősége
   */
  softwareDevContact: string;
  /**
   * ISO-3166 alpha2 country code of the software's developer
   *
   * A program fejlesztőjének ISO-3166 alpha2 országkódja
   */
  softwareDevCountryCode?: string;
  /**
   * Tax number of the software's developer
   *
   * A program fejlesztőjének adószáma
   */
  softwareDevTaxNumber?: string;
}